Improve config.go code and update documentation
The configuration source used magic numbers. Some of these have
now been added as constants.

The loading of configuration files had a `defer` in the for-loop. It is
removed and the functionality added as a new (private) method. We also
introduce the `newConfig` function, which is a helper, and added some
tests.

When the log level is changed, a message is now logged denoting this
event.

Documentation was outdated or simply not available for a few important
configuration (for example all the ``*-address` settings).