Add directive @skipMerge to skip types and interfaces
We have a schema of a service which provides a type. This is injected
through an SDK in other services, and provided in the Schema. However,
when merging, only the type or interface that the original service
providing it, should be considered. This is now possible using the
`@skipMerge` directive.